backend/app/requirements.txt
fastapi
uvicorn[standard]
celery[redis]
pillow
imagehash
sentence-transformers
scikit-learn
firebase-admin
python-dotenv
python-jose[cryptography]
